Change amount, fee and nonce back to NUMERIC(78) type

+++ db/migrations/000003_transaction_base.up.sql
@@ -2,9 +2,9 @@ CREATE TABLE "transaction_base" (
     tx_hash                BYTEA PRIMARY KEY,
     tx_type                SMALLINT NOT NULL,
     from_state_id          BIGINT NOT NULL,
-    amount                 BIGINT NOT NULL,
-    fee                    BIGINT NOT NULL,
-    nonce                  BIGINT NOT NULL,
+    amount                 NUMERIC(78) NOT NULL,
+    fee                    NUMERIC(78) NOT NULL,
+    nonce                  NUMERIC(78) NOT NULL,
     signature              BYTEA NOT NULL,
     included_in_commitment INTEGER REFERENCES commitment,
     error_message          TEXT
